requirements.txt
discord.py>=2.3.2
yt-dlp>=2024.12.23
PyNaCl>=1.5.0
pyinstaller>=6.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0 